# Remove stale moveaxis optimization in attention.